✅ HCJQP0 code: done
Task: 202604210859-HCJQP0
Primary: code
Agent: CODER
Author: CODER
Status: DONE
Comment: verified: split hosted-close PR command into precheck, execute, postcheck, report, and shared type modules | details: targeted CLI/command tests, typecheck, lint/format, and framework bootstrap passed.
