+++ packages/agentplane/src/commands/task/hosted-close-pr.report.ts
@@ -1,0 +1,73 @@
+import { createCliEmitter } from "../../cli/output.js";
+
+import { shortHostedCloseSha } from "./hosted-close-pr.precheck.js";
+import type {
+  HostedClosePrExecutionResult,
+  HostedClosePrNotice,
+  HostedClosePrOutcome,
+} from "./hosted-close-pr.types.js";
+
+function normalizeGithubPrLink(
+  prNumber: number,
+  prUrl: string | null,
+  verb: "linked to" | "created",
+): string {
+  return prUrl?.trim()
+    ? `${verb} GitHub PR #${prNumber}: ${prUrl.trim()}`
+    : `${verb} GitHub PR #${prNumber}`;
+}
+
+function reportNotice(output: ReturnType<typeof createCliEmitter>, notice: HostedClosePrNotice) {
+  if (notice.level === "warn") {
+    output.warn(notice.message);
+    return;
+  }
+  output.info(notice.message);
+}
+
+export function reportHostedClosePrOutcome(outcome: HostedClosePrOutcome): void {
+  reportHostedClosePrExecutionResult({ notices: [], outcome });
+}
+
+export function reportHostedClosePrExecutionResult(result: HostedClosePrExecutionResult): void {
+  const output = createCliEmitter();
+  for (const notice of result.notices) {
+    reportNotice(output, notice);
+  }
+
+  switch (result.outcome.kind) {
+    case "canonical-already-present": {
+      output.info(
+        `hosted-close-pr skipped: ${result.outcome.taskId} is already closed on ${result.outcome.baseBranch || "the base branch"} for merge ${shortHostedCloseSha(result.outcome.mergeCommit)}`,
+      );
+      return;
+    }
+    case "base-already-recorded": {
+      output.info(
+        `hosted close already recorded on ${result.outcome.baseBranch}; no follow-up PR needed`,
+      );
+      output.success(
+        "task hosted-close-pr",
+        result.outcome.taskId,
+        `hosted close already recorded on ${result.outcome.baseBranch}; skipped follow-up PR`,
+      );
+      return;
+    }
+    case "existing-pr": {
+      output.success(
+        "task hosted-close-pr",
+        result.outcome.taskId,
+        normalizeGithubPrLink(result.outcome.prNumber, result.outcome.prUrl, "linked to"),
+      );
+      return;
+    }
+    case "created-pr": {
+      output.success(
+        "task hosted-close-pr",
+        result.outcome.taskId,
+        normalizeGithubPrLink(result.outcome.prNumber, result.outcome.prUrl, "created"),
+      );
+      return;
+    }
+  }
+}
